// src/store/static/localization/types/LocalizationDataKey.ts
export enum LocalizationDataKey {
  APP_TITLE_TEXT = 'APP_TITLE_TEXT',
  TASK_LIST_ADD_TASK_BUTTON_TEXT = 'TASK_LIST_ADD_TASK_BUTTON_TEXT',
  TASK_LIST_ADD_TASK_BUTTON_TITLE = 'TASK_LIST_ADD_TASK_BUTTON_TITLE',
  TASK_LIST_EDIT_TASK_BUTTON_TITLE = 'TASK_LIST_EDIT_TASK_BUTTON_TITLE',
  TASK_LIST_REMOVE_TASK_BUTTON_TITLE = 'TASK_LIST_REMOVE_TASK_BUTTON_TITLE',
  TASK_LIST_CHANGE_TASK_COMPLETION_BUTTON_COMPLETED_STATE_TITLE = 'TASK_LIST_CHANGE_TASK_COMPLETION_BUTTON_COMPLETED_STATE_TITLE',
  TASK_LIST_CHANGE_TASK_COMPLETION_BUTTON_UNCOMPLETED_STATE_TITLE = 'TASK_LIST_CHANGE_TASK_COMPLETION_BUTTON_UNCOMPLETED_STATE_TITLE',
  TASK_LIST_REMOVE_ALL_TASKS_BUTTON_TEXT = 'TASK_LIST_REMOVE_ALL_TASKS_BUTTON_TEXT',
  TASK_LIST_REMOVE_ALL_TASKS_BUTTON_TITLE = 'TASK_LIST_REMOVE_ALL_TASKS_BUTTON_TITLE',
  TASK_LIST_REMOVE_ALL_COMPLETED_TASKS_BUTTON_TEXT = 'TASK_LIST_REMOVE_ALL_COMPLETED_TASKS_BUTTON_TEXT',
  TASK_LIST_REMOVE_ALL_COMPLETED_TASKS_BUTTON_TITLE = 'TASK_LIST_REMOVE_ALL_COMPLETED_TASKS_BUTTON_TITLE',
  TASK_LIST_TOTAL_COUNTER = 'TASK_LIST_TOTAL_COUNTER',
  TASK_LIST_NO_TASKS = 'TASK_LIST_NO_TASKS',
  TASK_FILTER_SWITCHER_ALL = 'TASK_FILTER_SWITCHER_ALL',
  TASK_FILTER_SWITCHER_ACTIVE = 'TASK_FILTER_SWITCHER_ACTIVE',
  TASK_FILTER_SWITCHER_COMPLETED = 'TASK_FILTER_SWITCHER_COMPLETED',
  POPUPS_CONFIRMATION_POPUP_CONFIRM_BUTTON_TEXT = 'POPUPS_CONFIRMATION_POPUP_CONFIRM_BUTTON_TEXT',
  POPUPS_CONFIRMATION_POPUP_CONFIRM_BUTTON_TITLE = 'POPUPS_CONFIRMATION_POPUP_CONFIRM_BUTTON_TITLE',
  POPUPS_CONFIRMATION_POPUP_CANCEL_BUTTON_TEXT = 'POPUPS_CONFIRMATION_POPUP_CANCEL_BUTTON_TEXT',
  POPUPS_CONFIRMATION_POPUP_CANCEL_BUTTON_TITLE = 'POPUPS_CONFIRMATION_POPUP_CANCEL_BUTTON_TITLE',
  POPUPS_DATA_MANAGEMENT_POPUP_SAVE_BUTTON_TEXT = 'POPUPS_DATA_MANAGEMENT_POPUP_SAVE_BUTTON_TEXT',
  POPUPS_DATA_MANAGEMENT_POPUP_SAVE_BUTTON_TITLE = 'POPUPS_DATA_MANAGEMENT_POPUP_SAVE_BUTTON_TITLE',
  POPUPS_DATA_MANAGEMENT_POPUP_CANCEL_BUTTON_TEXT = 'POPUPS_DATA_MANAGEMENT_POPUP_CANCEL_BUTTON_TEXT',
  POPUPS_DATA_MANAGEMENT_POPUP_CANCEL_BUTTON_TITLE = 'POPUPS_DATA_MANAGEMENT_POPUP_CANCEL_BUTTON_TITLE',
  POPUPS_REMOVE_TASK_CONFIRMATION_POPUP_TEXT = 'POPUPS_REMOVE_TASK_CONFIRMATION_POPUP_TEXT',
  POPUPS_REMOVE_TASK_CONFIRMATION_POPUP_TITLE = 'POPUPS_REMOVE_TASK_CONFIRMATION_POPUP_TITLE',
  POPUPS_REMOVE_ALL_TASKS_CONFIRMATION_POPUP_TEXT = 'POPUPS_REMOVE_ALL_TASKS_CONFIRMATION_POPUP_TEXT',
  POPUPS_REMOVE_ALL_TASKS_CONFIRMATION_POPUP_TITLE = 'POPUPS_REMOVE_ALL_TASKS_CONFIRMATION_POPUP_TITLE',
  POPUPS_REMOVE_ALL_COMPLETED_TASKS_CONFIRMATION_POPUP_TEXT = 'POPUPS_REMOVE_ALL_COMPLETED_TASKS_CONFIRMATION_POPUP_TEXT',
  POPUPS_REMOVE_ALL_COMPLETED_TASKS_CONFIRMATION_POPUP_TITLE = 'POPUPS_REMOVE_ALL_COMPLETED_TASKS_CONFIRMATION_POPUP_TITLE'
}